+++ Sources/BlueConnect/CentralManagerProxy/BleCentralManagerProxy.swift
@@ -362,6 +362,36 @@ extension BleCentralManagerProxy {
     
 }
 
+// MARK: - Peripherals retrieval
+
+extension BleCentralManagerProxy {
+    
+    /// Retrieves all peripherals that are connected to the system and implement any of the services listed in serviceUUIDs.
+    ///
+    /// This method returns peripherals that are currently connected to the system and offer specific services.
+    ///
+    /// - Parameters:
+    ///   - serviceUUIDs: A list of service UUIDs (represented by `CBUUID` objects) to filter connected peripherals.
+    ///
+    /// - Returns: A list of objects implementing the `BlePeripheral` protocol.
+    func retrieveConnectedPeripherals(withServices serviceUUIDs: [CBUUID]) -> [BlePeripheral] {
+        return centralManager.retrieveConnectedPeripherals(withServiceIds: serviceUUIDs)
+    }
+    
+    /// Retrieve all the peripherals with the corresponding identifiers.
+    ///
+    /// This method retrieves previously connected peripherals with specific identifiers.
+    ///
+    /// - Parameters:
+    ///   - identifiers: A list of peripheral identifiers (represented by `UUID` objects) from which peripheral objects can be retrieved.
+    ///
+    /// - Returns: A list of objects implementing the `BlePeripheral` protocol.
+    func retrievePeripherals(withIdentifiers identifiers: [UUID]) -> [BlePeripheral] {
+        return centralManager.retrievePeripherals(withIds: identifiers)
+    }
+    
+}
+
 // MARK: - State change
 
 extension BleCentralManagerProxy {