+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release Notes
 TBD
 
 - [#4](https://github.com/danielepantaleone/BlueConnect/pull/4) Added registry based subscription to avoid swift continuation misuse
+- [#5](https://github.com/danielepantaleone/BlueConnect/pull/5) Added `retrieveConnectedPeripherals` and `retrievePeripherals` methods to central manager proxy
 - Added support for RSSI read in peripheral proxy [538e3d8](https://github.com/danielepantaleone/BlueConnect/commit/538e3d81a835678c9b2a8aedb4e46e1c8f4599eb)
 
 **Full Changelog**: [1.1.0...1.2.0](https://github.com/danielepantaleone/BlueConnect/compare/1.1.0...1.2.0)